% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lookup_data.R
\name{add_para_metadata}
\alias{add_para_metadata}
\title{add_para_metadata}
\usage{
add_para_metadata(df, lookup_para_path, parameters_path)
}
\arguments{
\item{df}{df}

\item{lookup_para_path}{lookup_para_path}

\item{parameters_path}{parameters_path}
}
\value{
return "df" with added parameter metadata
}
\description{
add_para_metadata
}
