% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lookup_data.R
\name{add_site_metadata}
\alias{add_site_metadata}
\title{add_site_metadata}
\usage{
add_site_metadata(df, site_path)
}
\arguments{
\item{df}{df}

\item{site_path}{site_path}
}
\value{
data frame with added site metadata
}
\description{
add_site_metadata
}
