% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_defined}
\alias{all_defined}
\title{Check if all strings are not empty}
\usage{
all_defined(x)
}
\arguments{
\item{x}{vector of character}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check if all strings are not empty
}
