% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bwb_data.R
\name{cat_green_bold_0}
\alias{cat_green_bold_0}
\title{Helper function: cat_green_bold_0}
\usage{
cat_green_bold_0(...)
}
\arguments{
\item{...}{text passed to crayon::green()}
}
\value{
formatted text output
}
\description{
Helper function: cat_green_bold_0
}
