% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bwb_data.R
\name{cat_red_bold_0}
\alias{cat_red_bold_0}
\title{Helper function: cat_red_bold_0}
\usage{
cat_red_bold_0(...)
}
\arguments{
\item{...}{text passed to crayon::red}
}
\value{
formatted text output
}
\description{
Helper function: cat_red_bold_0
}
