% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{column_pattern_gather_ignore}
\alias{column_pattern_gather_ignore}
\title{Helper function: column_pattern_gather_ignore}
\usage{
column_pattern_gather_ignore(
  fields = c("Datum", "KN", "[iI]nterne Nr.", "Name der", "Ort", "Probe", "Prü",
    "Untersuchung", "Labor", "Jahr", "Galer", "Detail", "Meß", "Zeit", "Bezei", "Monat")
)
}
\arguments{
\item{fields}{column names to be ignored for gathering (default: c(
"Datum", "KN", "[iI]nterne Nr.", "Name der", "Ort", "Probe", "Pr\\u00FC",
"Untersuchung", "Labor", "Jahr", "Galer", "Detail", "Me\\u00DF", "Zeit",
"Bezei", "Monat")}
}
\value{
vector with ignored columns for gathering
}
\description{
Helper function: column_pattern_gather_ignore
}
