% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{column_pattern_gather_ignore_clean}
\alias{column_pattern_gather_ignore_clean}
\title{Helper function: column_pattern_gather_ignore_clean}
\usage{
column_pattern_gather_ignore_clean(
  fields = c("LabSampleCode", "Date", "Time", "Waterbody", "ExSiteCode", "Site")
)
}
\arguments{
\item{fields}{column names to be ignored for gathering (default:
c("LabSampleCode", "Date", "Time", "Waterbody", "ExSiteCode", Site")}
}
\value{
vector with ignored columns for gathering
}
\description{
Helper function: column_pattern_gather_ignore_clean
}
