% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_phreeqc_input.R
\name{convert_phreeqc_input_to_wide}
\alias{convert_phreeqc_input_to_wide}
\title{Convert PhreeqC input to "wide" format}
\usage{
convert_phreeqc_input_to_wide(phreeqc_input)
}
\arguments{
\item{phreeqc_input}{PhreeqC input as retrieved by \code{\link{get_phreeqc_data}}}
}
\value{
PhreeqC input in "wide" format
}
\description{
Convert PhreeqC input to "wide" format
}
