% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_sf.R
\name{convert_to_sf}
\alias{convert_to_sf}
\title{Convert to SF}
\usage{
convert_to_sf(
  df,
  crs_source = 25833,
  crs_target = 4326,
  col_coord_x = "Rechtswert_UTM_33_N",
  col_cood_y = "Hochwert_UTM_33_N"
)
}
\arguments{
\item{df}{data frame or tibble with spatial data}

\item{crs_source}{origingal CRS (default: 25833)}

\item{crs_target}{target CRS (default: 4326)}

\item{col_coord_x}{column name of latitude (default: "Rechtswert_UTM_33_N")}

\item{col_cood_y}{column name of longitude (default: "Hochwert_UTM_33_N")}
}
\value{
data frame or tibble converted to sf
}
\description{
Convert to SF
}
\examples{
gwl_master <- jsonlite::fromJSON("https://kwb-r.github.io/wasserportal/stations_gwl_master.json")
convert_to_sf(gwl_master)

}
