% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_xlsx_files.R
\name{copy_xlsx_files}
\alias{copy_xlsx_files}
\title{Helper function: copy_xlsx_files}
\usage{
copy_xlsx_files(
  from_dir,
  to_dir,
  overwrite = FALSE,
  recursive = TRUE,
  file_pattern = "[xX][lL][sS][xX]"
)
}
\arguments{
\item{from_dir}{input directory with xlsx files}

\item{to_dir}{target directory where to copy the xlsx files}

\item{overwrite}{should existing files be overwritten (TRUE) otherwise (FALSE)
? (default: FALSE)}

\item{recursive}{if TRUE recursively find all xlsx files in the directory
specified in parameter "from_dir" (default: TRUE)}

\item{file_pattern}{pattern for identifying xlsx fles (default: "[xX][lL][sS][xX]")}
}
\description{
Helper function: copy_xlsx_files
}
