% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emshoff91.R
\name{create_emshoff91_import}
\alias{create_emshoff91_import}
\title{Create EMSHOFF91 Import Data Frame}
\usage{
create_emshoff91_import(
  ods_dir,
  files_to_ignore = c("cl25", "clliste", "gf_gm", "gwnguete", "rupelauf", "salzlast")
)
}
\arguments{
\item{ods_dir}{directory to ".ods" files created manually by importing original
".wq1" files into LibreOffice 7.0 on Ubuntu with encoding (Western Europe
(DOS/OS2-437/US)) and exporting to ".ods" format}

\item{files_to_ignore}{tidied names of files to ignore due to complex data input
structure not yet covered by importer (default: c("cl25", "clliste", "gf_gm", 
"gwnguete", "rupelauf", "salzlast"))}
}
\value{
data frame with columns "ods_paths" (full paths to ".ods" files),
"ods_files" (their "basenames") and "ods_names_clean" (tidied names used as
identifier)
}
\description{
Create EMSHOFF91 Import Data Frame
}
\examples{
\dontrun{
ods_dir <- "<replace-with-path-to-files>/emshoff91/converted_ods"
emshoff91_import <- create_emshoff91_import(ods_dir)
}
}
