% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement-chains.R
\name{create_sftp_connection}
\alias{create_sftp_connection}
\title{Measurement Chains: Create SFTP credentials (curl backend)}
\usage{
create_sftp_connection()
}
\value{
list with server/username/password/port
}
\description{
Measurement Chains: Create SFTP credentials (curl backend)
}
