% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement-chains.R
\name{download_measurementchains_data}
\alias{download_measurementchains_data}
\title{Measurement Chains: download data (SFTP via curl)}
\usage{
download_measurementchains_data(
  sftp_paths,
  target_directory = temp_dir(),
  run_parallel = TRUE,
  debug = FALSE
)
}
\arguments{
\item{sftp_paths}{character vector with relative paths to files (under MESSKETTEN_REMOTE_DIR)}

\item{target_directory}{target directory}

\item{run_parallel}{default: TRUE (will create separate curl handles; ok)}

\item{debug}{show debug messages (default: FALSE)}
}
\value{
character vector of local file paths
}
\description{
Measurement Chains: download data (SFTP via curl)
}
