% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emshoff91_remap_values.R
\name{emshoff91_list_to_df}
\alias{emshoff91_list_to_df}
\title{Emshoff 91: list to data frame}
\usage{
emshoff91_list_to_df(emshoff91_list)
}
\arguments{
\item{emshoff91_list}{list as retrieved by \code{\link{read_multiple_emshoff91_ods}}}
}
\value{
tibble
}
\description{
Emshoff 91: list to data frame
}
