% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bwb_data.R
\name{gather_and_join_1}
\alias{gather_and_join_1}
\title{Helper function: gather_and_join_1}
\usage{
gather_and_join_1(tmp_data, columns_keep, metadata, dbg = FALSE)
}
\arguments{
\item{tmp_data}{tmp_data}

\item{columns_keep}{columns_keep}

\item{metadata}{metadata}

\item{dbg}{dbg (default: FALSE)}
}
\value{
gathered and joined data frame
}
\description{
Helper function: gather_and_join_1
}
