% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bwb_data.R
\name{gather_and_join_2}
\alias{gather_and_join_2}
\title{Helper function: gather_and_join_2}
\usage{
gather_and_join_2(tmp_content, columns_keep, header)
}
\arguments{
\item{tmp_content}{tmp_content}

\item{columns_keep}{columns_keep}

\item{header}{header}
}
\value{
gathered and joined data frame
}
\description{
Helper function: gather_and_join_2
}
