% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bwb_data.R
\name{gather_ignore}
\alias{gather_ignore}
\title{Helper function: gather_ignore}
\usage{
gather_ignore(fields = c("Datum", "KN", "[iI]nterne Nr.", "Name der",
  "Ort", "Probe", "Pr\\\\u00fcf", "Untersuchung", "Labor", "Jahr", "Galer",
  "Detail", "Me\\\\u00DF", "Zeit", "Bezei", "Monat"))
}
\arguments{
\item{fields}{column names to be ignored for gathering (default: c(
"Datum", "KN", "[iI]nterne Nr.", "Name der", "Ort", "Probe", "Pr\\u00fcf",
"Untersuchung", "Labor", "Jahr", "Galer", "Detail", "Me\\u00DF", "Zeit",
"Bezei", "Monat")}
}
\value{
vector with ignored columns for gathering
}
\description{
Helper function: gather_ignore
}
