% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bwb_data.R
\name{gather_ignore_clean}
\alias{gather_ignore_clean}
\title{Helper function: gather_ignore_clean}
\usage{
gather_ignore_clean(fields = c("LabSampleCode", "Date", "Time", "Waterbody",
  "ExSiteCode", "Site"))
}
\arguments{
\item{fields}{column names to be ignored for gathering (default: 
c("LabSampleCode", "Date", "Time", "Waterbody", "ExSiteCode", Site")}
}
\value{
vector with ignored columns for gathering
}
\description{
Helper function: gather_ignore_clean
}
