% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_foerdermengen.R
\name{get_foerdermengen}
\alias{get_foerdermengen}
\title{get_foerdermengen}
\usage{
get_foerdermengen(
  xlsx_path,
  sheet_name = "WW Q Rhow ",
  sheet_range = "A4:S127"
)
}
\arguments{
\item{xlsx_path}{path to xlsx file with pumping rates}

\item{sheet_name}{sheet_name (default: "WW Q Rhow ")}

\item{sheet_range}{sheet_range  (default: "A4:S127")}
}
\value{
data frame with annual pumping rates per waterworks
}
\description{
get_foerdermengen
}
