% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_foerdermengen_gal_fri.R
\name{get_foerdermengen_gal_fri}
\alias{get_foerdermengen_gal_fri}
\title{Get Abtraction of Friedrichshagen Well Galleries}
\usage{
get_foerdermengen_gal_fri(path)
}
\arguments{
\item{path}{path to "2018-04-27 Rohwasser Bericht - Galeriefördermengen.xlsx"}
}
\value{
tidy data frame with abstraction rates for waterworks Friedrichshagen
}
\description{
Get Abtraction of Friedrichshagen Well Galleries
}
