% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measurement_chain_data_on_cloud.R
\name{get_measurement_chain_data_on_cloud}
\alias{get_measurement_chain_data_on_cloud}
\title{Get Measurement Chain Data on KWB Cloud}
\usage{
get_measurement_chain_data_on_cloud()
}
\value{
data frame with the content of "mc_data.zip" in the GeoSalz project
  folder on the KWB nextcloud. The SFTP paths to the files from which the
  data in "mc_data.zip" originate are returned in attribute "sftp_paths".
}
\description{
Get Measurement Chain Data on KWB Cloud
}
