% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement-chains.R
\name{get_measurementchains_metadata}
\alias{get_measurementchains_metadata}
\title{Measurement Chains: Get Metadata}
\usage{
get_measurementchains_metadata(file = extdata_file("metadata_messketten.csv"))
}
\arguments{
\item{file}{path to measurement chains metadata file. Default:
kwb.geosalz:::extdata_file("metadata_messketten.csv")}
}
\value{
tibble with measurement chains metadata
}
\description{
Measurement Chains: Get Metadata
}
\examples{
mc_metadata <- kwb.geosalz::get_measurementchains_metadata()
str(mc_metadata)
mc_metadata
}
