% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measurmentchains_data_stats.R
\name{get_measurmentchains_data_stats}
\alias{get_measurmentchains_data_stats}
\title{Measurement Chains: get statistics for data}
\usage{
get_measurmentchains_data_stats(mc_data)
}
\arguments{
\item{mc_data}{tibble with measurement chains data as retrieved by 
\code{\link{read_measurementchains_data}}}
}
\value{
tibble with colunns datetime min/max, q10 (10% percentile), mean, 
median, q90 (90% percentile), max
}
\description{
Measurement Chains: get statistics for data
}
