% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_meta_sheet_or_stop}
\alias{get_meta_sheet_or_stop}
\title{Helper function: get_meta_sheet_or_stop}
\usage{
get_meta_sheet_or_stop(sheets, pattern, file)
}
\arguments{
\item{sheets}{sheets}

\item{pattern}{pattern}

\item{file}{file}
}
\value{
meta sheet name
}
\description{
Helper function: get_meta_sheet_or_stop
}
