% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoringchains.R
\name{get_monitoringchains_files}
\alias{get_monitoringchains_files}
\title{Monitoring Chains: Get Tidied Files Metadata}
\usage{
get_monitoringchains_files(
  sftp_connection = create_sftp_connection(),
  debug = FALSE
)
}
\arguments{
\item{sftp_connection}{an SFTP connnection as retrieved by 
\code{\link{create_sftp_connection}}}

\item{debug}{show debug messages (default: FALSE)}
}
\value{
tibble with information on available files and tidied meta-information 
based on file naming
}
\description{
Monitoring Chains: Get Tidied Files Metadata
}
\examples{
\dontrun{
mc_files <- kwb.geosalz::get_monitoringchains_files()
str(mc_files)
}
}
