% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoringchains.R
\name{get_monitoringchains_metadata}
\alias{get_monitoringchains_metadata}
\title{Monitoring Chains: Get Metadata}
\usage{
get_monitoringchains_metadata(
  file = system.file("extdata/metadata_messketten.csv", package = "kwb.geosalz")
)
}
\arguments{
\item{file}{path to monitoring chains metadata file (default: 
system.file("extdata/metadata_messketten.csv", package = "kwb.geosalz"))}
}
\value{
tibble with monitoring chains metadata
}
\description{
Monitoring Chains: Get Metadata
}
\examples{
mc_metadata <- kwb.geosalz::get_monitoringchains_metadata()
str(mc_metdata)
}
