% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pandoc_info.R
\name{get_pandoc_info}
\alias{get_pandoc_info}
\title{Get Information on Pandoc}
\usage{
get_pandoc_info()
}
\value{
data frame with columns \code{pandoc_directory},
  \code{pandoc_version} if Pandoc is installed, otherwise a message is
  printed that pandoc is not installed.
}
\description{
Get Information on Pandoc
}
