% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lookup_data.R
\name{get_parameters_meta}
\alias{get_parameters_meta}
\title{get_parameters_meta}
\usage{
get_parameters_meta(xlsx_path, sheet_name = "nur Parameterliste")
}
\arguments{
\item{xlsx_path}{path to EXCEL spreadsheet with parameter metadata}

\item{sheet_name}{name of sheet containing metadata table
(default: "nur Parameterliste")}
}
\value{
imported parameter metadata with cleaned columns names
}
\description{
get_parameters_meta
}
