% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_phreeqc_input.R
\name{get_phreeqc_data}
\alias{get_phreeqc_data}
\title{Get PhreeqC data}
\usage{
get_phreeqc_data(lab_bwb)
}
\arguments{
\item{lab_bwb}{imported BWB lab data as retrieved by \code{\link{read_lab_bwb}}}
}
\value{
tibble with columns solution, par_name_phreeqc (which are not empty or 
NA) and numeric_value
}
\description{
Get PhreeqC data
}
