% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bwb_data.R
\name{get_site_id}
\alias{get_site_id}
\title{Helper function: get_site_id}
\usage{
get_site_id(string, pattern = "^[0-9]{1,4}")
}
\arguments{
\item{string}{vector with character strings}

\item{pattern}{pattern used for identifying site_id (default: "^[0-9]{1,4}")}
}
\value{
extracted site_id`s from input string
}
\description{
Helper function: get_site_id
}
