% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bwb_data.R
\name{import_labor}
\alias{import_labor}
\title{import_labor}
\usage{
import_labor(files, export_dir, func = read_bwb_header2)
}
\arguments{
\item{files}{vector with full paths of xlsx input files}

\item{export_dir}{export directory}

\item{func}{function to be used (default: \code{read_bwb_header2})}
}
\value{
list with length equal to number of input files
}
\description{
import_labor
}
