% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement-chains.R
\name{order_measurement_chain_data}
\alias{order_measurement_chain_data}
\title{Order Measurement Chain Data}
\usage{
order_measurement_chain_data(data)
}
\arguments{
\item{data}{data frame as retrieved by
\code{\link{read_measurementchains_data}}}
}
\value{
\code{data}, ordered by "parameter", "sensor_id", "datum_uhrzeit"
}
\description{
Order Measurement Chain Data
}
