% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_measurementchain_and_welloperation.R
\name{plot_measurementchain_and_well_operation}
\alias{plot_measurementchain_and_well_operation}
\title{Plot measurementchain and well operation in combined plot}
\usage{
plot_measurementchain_and_well_operation(
  mc_dat,
  well_op_data_meta,
  brunnen_nr = 9,
  para = "Leitfaehigkeit",
  y_label = "elektr. Leitfähigkeit (µS/cm)",
  date_min = as.Date("2023-05-10"),
  date_max = Sys.Date()
)
}
\arguments{
\item{mc_dat}{mc_dat}

\item{well_op_data_meta}{well_op_data_meta}

\item{brunnen_nr}{well id (default: 9)}

\item{para}{parameter (either: "Leitfaehigkeit" or "Temperatur")}

\item{y_label}{y label (default: "elektr. Leitfaehigkeit (µS/cm)")}

\item{date_min}{minimum date for plotting (default: as.Date("2023-05-10"))}

\item{date_max}{maximum date for plotting (default: Sys.Date())}
}
\value{
combined plot
}
\description{
Plot measurementchain and well operation in combined plot
}
