% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_measurementchains.R
\name{plot_measurementchains}
\alias{plot_measurementchains}
\title{Measurement Chains: plott}
\usage{
plot_measurementchains(mc_data, para = "Leitfaehigkeit")
}
\arguments{
\item{mc_data}{as retrieved by \code{\link{read_measurementchains_data}}}

\item{para}{parameter to plot "Leitfaehigkeit" or "Temperatur" (default: 
Leitfaehigkeit")}
}
\value{
plot of selected measurement chain parameter
}
\description{
Measurement Chains: plott
}
