% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_phreeqc_input.R
\name{prepare_phreeqc_input}
\alias{prepare_phreeqc_input}
\title{Prepare PhreeqC input}
\usage{
prepare_phreeqc_input(lab_bwb_phreeqc, title = "Test Dataset")
}
\arguments{
\item{lab_bwb_phreeqc}{selected BWB lab data as retrieved by \code{\link{get_phreeqc_data}}}

\item{title}{user defined title (default: "Test Dataset")}
}
\value{
data frame with input structure for kwb.phreeqc
}
\description{
Prepare PhreeqC input
}
