% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{print_datatype_info_if}
\alias{print_datatype_info_if}
\title{Helper function: print_datatype_info_if}
\usage{
print_datatype_info_if(dbg, tbl_datatypes, columns_keep)
}
\arguments{
\item{dbg}{dbg}

\item{tbl_datatypes}{tbl_datatypes}

\item{columns_keep}{columns_keep}
}
\description{
Helper function: print_datatype_info_if
}
