% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bwb_data.R
\name{read_bwb_data}
\alias{read_bwb_data}
\title{Import: read_bwb_data}
\usage{
read_bwb_data(
  files,
  meta_pattern = "META",
  keep_pattern = column_pattern_gather_ignore(),
  site_id_pattern = "^[0-9]{1,4}",
  dbg = TRUE
)
}
\arguments{
\item{files}{file path(s) to EXCEL spreadsheet}

\item{meta_pattern}{(default: "META")}

\item{keep_pattern}{(default: \code{column_pattern_gather_ignore})}

\item{site_id_pattern}{(default: "^[0-9]{1,4}")}

\item{dbg}{debug (default: TRUE)}
}
\value{
data.table with imported xls(x) files
}
\description{
wrapper around \code{read_bwb_header2} and
\code{read_bwb_header1_meta}
}
