% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bwb_data.R
\name{read_bwb_header1_meta}
\alias{read_bwb_header1_meta}
\title{Import: read_bwb_header1_meta}
\usage{
read_bwb_header1_meta(
  file,
  meta_pattern = "META",
  keep_pattern = column_pattern_gather_ignore(),
  dbg = FALSE
)
}
\arguments{
\item{file}{path(s) to EXCEL spreadsheet}

\item{meta_pattern}{meta_pattern default("META")}

\item{keep_pattern}{keep_pattern (default: \code{column_pattern_gather_ignore})}

\item{dbg}{debug (default: FALSE)}
}
\value{
data.table with imported xls(x) files
}
\description{
Import: read_bwb_header1_meta
}
