% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bwb_data.R
\name{read_bwb_header4}
\alias{read_bwb_header4}
\title{Import: read_bwb_header4}
\usage{
read_bwb_header4(
  file,
  skip = 4,
  keep_pattern = gather_ignore(),
  site_id_pattern = "^[0-9]{1,4}",
  dbg = TRUE
)
}
\arguments{
\item{file}{file path(s) to EXCEL spreadsheet}

\item{skip}{number of rows to skip in each sheet (default: 4)}

\item{keep_pattern}{(default: gather_ignore())}

\item{site_id_pattern}{(default: "^[0-9]{1,4}")}

\item{dbg}{debug (default: TRUE)}
}
\value{
data.table with imported xls(x) files
}
\description{
Import: read_bwb_header4
}
