% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emshoff91.R
\name{read_emshoff91_ods}
\alias{read_emshoff91_ods}
\title{Reads a Single EMSHOFF 91 ODS File into Tibble}
\usage{
read_emshoff91_ods(emshoff91_import_selected)
}
\arguments{
\item{emshoff91_import_selected}{a row as retrieved by \code{\link{create_emshoff91_import}}}
}
\value{
imports ods file into tibble
}
\description{
Reads a Single EMSHOFF 91 ODS File into Tibble
}
