% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isotopes.R
\name{read_isotopes}
\alias{read_isotopes}
\title{Read Isotopes}
\usage{
read_isotopes(path)
}
\arguments{
\item{path}{path to Isotopes delim (field separator ";")}
}
\value{
imported isotopes data
}
\description{
Read Isotopes
}
