% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lab_bwb.R
\name{read_lab_bwb}
\alias{read_lab_bwb}
\title{Read Lab BWB}
\usage{
read_lab_bwb(path, sheet = "Analysen")
}
\arguments{
\item{path}{path to file with lab BWB data}

\item{sheet}{name of sheet containing analysis data. Default: "Analysen"}
}
\value{
cleaned data frame with master data and lab values for all samples but 
only for selected parameters (columns A-BA and HB-HC)
}
\description{
Read Lab BWB
}
