% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lab_bwb.R
\name{read_master_data}
\alias{read_master_data}
\title{Read Master Data}
\usage{
read_master_data(path)
}
\arguments{
\item{path}{path to file with master data (currently in file: lab BWB data)}
}
\value{
imported master data contained in sheet "Stammdaten"
}
\description{
Read Master Data
}
