% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement-chains.R
\name{read_measurementchains_data}
\alias{read_measurementchains_data}
\title{Measurement Chains: read csv data from multiple files}
\usage{
read_measurementchains_data(
  csv_files,
  datetime_installation = as_gmt_plus_one("2022-09-27 11:00:00"),
  run_parallel = TRUE,
  debug = FALSE
)
}
\arguments{
\item{csv_files}{vector of paths as retrieved by
\code{\link{download_measurementchains_data}}}

\item{datetime_installation}{datetime of first logger installation in well K10. 
Used to filter out older measurement data! Default: 
kwb.geosalz:::as_gmt_plus_one("2022-09-27 11:00:00")}

\item{run_parallel}{default: TRUE}

\item{debug}{show debug messages (default: FALSE)}
}
\value{
data frame with imported data from csv files
}
\description{
Measurement Chains: read csv data from multiple files
}
\examples{
\dontrun{
mc_files <- kwb.geosalz::get_measurementchains_files()
target_directory <- tempdir()
csv_files <- kwb.geosalz::download_measurementchains_data(
  sftp_paths = mc_files$sftp_path,
  target_directory
)
mc_data <- kwb.geosalz::read_measurementchains_data(csv_files)
}
}
