% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoringchains.R
\name{read_monitoringchains_data}
\alias{read_monitoringchains_data}
\title{Monitoring Chains: read csv data}
\usage{
read_monitoringchains_data(csv_paths, debug = FALSE)
}
\arguments{
\item{csv_paths}{character vector with paths to downloaded csv files. As 
retrieved by \code{\link{download_monitoringchains_data}}}

\item{debug}{show debug messages (default: FALSE)}
}
\value{
data frame with imported data from csv files
}
\description{
Monitoring Chains: read csv data
}
\examples{
\dontrun{
mc_files <- kwb.geosalz::get_monitoringchains_files()
target_directoy <- tempdir()
csv_paths <- download_monitoringchains_data(sftp_paths = mc_files$sftp_path,
target_directory)
mc_data <- read_monitoringchains_data(csv_paths)

}
}
