% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emshoff91.R
\name{read_multiple_emshoff91_ods}
\alias{read_multiple_emshoff91_ods}
\title{Reads Multiple EMSHOFF 91 ODS Files into List}
\usage{
read_multiple_emshoff91_ods(emshoff91_import)
}
\arguments{
\item{emshoff91_import}{a tibble as retrieved by \link{\code{create_emshoff91_import}}}
}
\value{
imports mulitple ods files into a list of tibbles
}
\description{
Reads Multiple EMSHOFF 91 ODS Files into List
}
\examples{
\dontrun{
ods_dir <- "<replace-with-path-to-files>/emshoff91/emshoff91/converted_ods"
ods_dir <- "C:/users/mrustl/Downloads/emshoff91/emshoff91/converted_ods"
emshoff91_import <- create_emshoff91_import(ods_dir)
read_multiple_emshoff91_ods(emshoff91_import)
}
}
