% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lab_bwb.R
\name{replace_nanb_with_na}
\alias{replace_nanb_with_na}
\title{Replace n.a. (not available) and n.b. (not determined) from lab data with NA}
\usage{
replace_nanb_with_na(string)
}
\arguments{
\item{string}{string}
}
\value{
string with NA instead of "n.a." or "n.b." (including 0-10 spaces between 
"n" and "a"/"b"
}
\description{
Replace n.a. (not available) and n.b. (not determined) from lab data with NA
}
\examples{
string <- c("19.2", "n.b.", "n. b.", "n.  b.", "n.a.", "n. a.", "n.  a.")
replace_nanb_with_na(string)
}
