% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bwb_data.R
\name{stop_formatted}
\alias{stop_formatted}
\title{Helper function: stop_formatted}
\usage{
stop_formatted(fmt, ...)
}
\arguments{
\item{fmt}{fmt}

\item{...}{arguments passed to fmt}
}
\value{
calls stop with fmt and additional arguments defined in ...
}
\description{
Helper function: stop_formatted
}
