% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_if_duplicated_samples_found.R
\name{stop_if_duplicated_samples_found}
\alias{stop_if_duplicated_samples_found}
\title{Helper function: stop if duplicated sample ids are found}
\usage{
stop_if_duplicated_samples_found(df, col_sampleid, path, sheet = "")
}
\arguments{
\item{df}{data frame with samples in wide format}

\item{col_sampleid}{column name of sample id}

\item{sheet}{optional in case EXCEL is used (default: "")}
}
\value{
error in case duplicated samples were found
}
\description{
Helper function: stop if duplicated sample ids are found
}
