% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bwb_data.R
\name{stop_on_missing_or_inform_on_extra_sheets}
\alias{stop_on_missing_or_inform_on_extra_sheets}
\title{Helper function: stop_on_missing_or_inform_on_extra_sheets}
\usage{
stop_on_missing_or_inform_on_extra_sheets(has_site_id, file, sheets)
}
\arguments{
\item{has_site_id}{has_site_id}

\item{file}{file}

\item{sheets}{sheets}
}
\description{
Helper function: stop_on_missing_or_inform_on_extra_sheets
}
