% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bwb_data.R
\name{to_full_metadata_4}
\alias{to_full_metadata_4}
\title{Helper function: to_full_metadata_4}
\usage{
to_full_metadata_4(header, file, sheet)
}
\arguments{
\item{header}{header}

\item{file}{file}

\item{sheet}{sheet}
}
\value{
data frame with metadata for header4 (EXCEL) files
}
\description{
Helper function: to_full_metadata_4
}
