% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_measurementchains_data.R
\name{write_measurementchains_data}
\alias{write_measurementchains_data}
\title{Measurement Chains: write csv data}
\usage{
write_measurementchains_data(
  mc_data,
  target_directory,
  to_zip = FALSE,
  debug = FALSE
)
}
\arguments{
\item{mc_data}{measurement chains data as retrieved by 
\code{\link{read_measurementchains_data}}}

\item{target_directory}{target directory}

\item{to_zip}{should data be zipped? (default: FALSE), if TRUE only a 
temporary csv file is created which will be subsequently zipped and deleted}

\item{debug}{print debug messages (default: FALSE)}
}
\value{
writes csv data to path
}
\description{
Measurement Chains: write csv data
}
