% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement-chains.R
\name{create_sftp_connection}
\alias{create_sftp_connection}
\title{Measurement Chains: Create an SFTP Connection}
\usage{
create_sftp_connection()
}
\value{
sftp connection
}
\description{
Measurement Chains: Create an SFTP Connection
}
