% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoringchains.R
\name{create_sftp_connection}
\alias{create_sftp_connection}
\title{Monitoring Chains: Create an SFTP Connection}
\usage{
create_sftp_connection()
}
\value{
sftp connection
}
\description{
Monitoring Chains: Create an SFTP Connection
}
