% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emshoff91_remap_values.R
\name{emshoff91_remap_values}
\alias{emshoff91_remap_values}
\title{Emshoff 91: remap values from imported tibble}
\usage{
emshoff91_remap_values(
  emshoff91_df,
  remap_list = list(fi_mi = "fi_mi_m_nn", ku_sto = "kupp_st", lf = "el_lf", strat =
    "stratigr", uv254 = "uv_ext")
)
}
\arguments{
\item{emshoff91_df}{tibble as retrieved by \code{\link{emshoff91_list_to_df}}}

\item{remap_list}{list with values to be remapped. Names of the list are columns 
values contained in list values should be mapped to (default: list(fi_mi = "fi_mi_m_nn", 
ku_sto = "kupp_st", lf = "el_lf", strat = "stratigr", uv254 = "uv_ext")),}
}
\value{
data frame with remapped values and deleted columns were this values 
were copied from
}
\description{
Emshoff 91: remap values from imported tibble
}
