% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoringchains.R
\name{get_measurementchains_metadata}
\alias{get_measurementchains_metadata}
\title{Measurement Chains: Get Metadata}
\usage{
get_measurementchains_metadata(
  file = system.file("extdata/metadata_messketten.csv", package = "kwb.geosalz")
)
}
\arguments{
\item{file}{path to measurement chains metadata file (default: 
system.file("extdata/metadata_messketten.csv", package = "kwb.geosalz"))}
}
\value{
tibble with measurement chains metadata
}
\description{
Measurement Chains: Get Metadata
}
\examples{
mc_metadata <- kwb.geosalz::get_measurementchains_metadata()
str(mc_metadata)
}
